LPS-72276 toCalendarBookingJSONObject() serializes the last instance recurrence.
The front-end should not deal with specific recurrences of calendar bookings
anymore, but only with the actual recurrence, that is the one that generates
the last instance. For this reason, we are sending the last instance recurrence
to the front end from now on. It may be slightly confusing but we bet it would
be yet more confusing to send two recurrences and have other developers use the
wrong one.