LPS-67210 Make day view move red line based on already retrieved current time.
The day view by default call scheduler.get('currentTimeFn'). In our case, it
would result in requesting the current time to the server (a) by the scheduler
and then (b) by the view. We can optimize this behavior by reusing the value
retrieved by the scheduler.

This is another change that would be better off in Alloy UI.